+++ Storage/MySQL/ImageMapper.php
@@ -21,7 +21,7 @@ final class ImageMapper extends AbstractMapper implements ImageMapperInterface
      */
     public static function getTableName()
     {
-        return 'bono_module_shop_product_images';
+        return self::getWithPrefix('bono_module_shop_product_images');
     }
 
     /**